requirements.txt
google-genai>=1.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.0.0
streamlit>=1.32.0
requests>=2.31.0
python-dotenv>=1.0.0
 